Allow setting of SSH_AUTH_SOCK via a global config
bbackup now supports a "global" (i.e. shared between the config
subdirectories) config which can specify (as so far the only
valid option) a value for the key "ssh-auth-sock-script-path".
This provides the path to a shell script which sets SSH_AUTH_SOCK,
from which bbackup extracts the value and passes it on to the borg
invocations as an environment variable.